# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseDetail": {
        "additionalProperties": false,
        "description": "`GET /cases/{case_id}`: the whole case, as architecture §24's detail screen shows it.",
        "properties": {
          "case": {
            "$ref": "#/components/schemas/CaseSummary"
          },
          "data_gaps": {
            "items": {
              "$ref": "#/components/schemas/DataGap"
            },
            "title": "Data Gaps",
            "type": "array"
          },
          "evaluation_no": {
            "title": "Evaluation No",
            "type": "integer"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteItem"
            },
            "title": "Notes",
            "type": "array"
          },
          "notifications": {
            "items": {
              "$ref": "#/components/schemas/NotificationItem"
            },
            "title": "Notifications",
            "type": "array"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "urgent_events": {
            "items": {
              "$ref": "#/components/schemas/UrgentEvent"
            },
            "title": "Urgent Events",
            "type": "array"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "case",
          "evaluation_no"
        ],
        "title": "CaseDetail",
        "type": "object"
      },
...
      "CaseSource": {
        "enum": [
          "offense",
          "hunt",
          "group"
        ],
        "title": "CaseSource",
        "type": "string"
      },
      "CaseStatus": {
        "description": "`cases.status`.",
        "enum": [
          "running",
          "decided",
          "no_ai_decision",
          "closed"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
...
      "CaseVerdict": {
        "enum": [
          "tp",
          "fp",
          "suspicious"
        ],
        "title": "CaseVerdict",
        "type": "string"
      },
...
      "FeedbackAnswer": {
        "additionalProperties": false,
        "description": "`POST /cases/{case_id}/feedback`: the stored feedback, with who wrote it.",
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/FeedbackReason"
          },
          "user_subject": {
            "title": "User Subject",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/CaseVerdict"
          }
        },
        "required": [
          "case_id",
          "user_subject",
          "verdict",
          "reason",
          "created_at"
        ],
        "title": "FeedbackAnswer",
        "type": "object"
      },
      "FeedbackReason": {
        "enum": [
          "correct",
          "was_tp_not_fp",
          "was_fp_not_tp",
          "missing_context",
          "wrong_urgent_events",
          "other"
        ],
        "title": "FeedbackReason",
        "type": "string"
      },
...
      "Level": {
        "description": "Severity level. Values compare as strings, not by severity.",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Level",
        "type": "string"
      },
...
      "Page_CaseSummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[CaseSummary]",
        "type": "object"
      },
...
      "Problem": {
        "description": "The shape of an RFC 9457 answer, for the UI's own error handling.\n\nNot a response model of any route: `ais0c_api.openapi` makes it every operation's `default`\nresponse in the schema. `errors`, `domains` and `list_names` appear only on the problems that\ncarry them.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "domains": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Domains"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "list_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "List Names"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The analyst UI's API. The UI talks only to this service.\n\nErrors are RFC 9457 problems whose `title` is a machine-readable code (`catalog.rule_not_found`,\n`qa.already_resolved`); the Turkish message the user reads is built by the UI from that code.\nTimes are ISO 8601 in UTC. Lists are cursor-paged with `?cursor=&limit=`.\n",
    "title": "ais0c analyst API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cases": {
      "get": {
        "description": "The offense queue, newest first (architecture §24).\n\n`from` and `to` bound `created_at`, `from` inclusive and `to` exclusive.",
        "operationId": "get_cases_api_v1_cases_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/CaseStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "notify_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/Level"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notify Level"
            }
          },
          {
            "in": "query",
            "name": "verdict",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/CaseVerdict"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/CaseSource"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "rule_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule Id"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CaseSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Cases",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "description": "The whole case: the report, the urgent events by rank, the recommendations, the\nVerification, the data gaps, the evidence, the notes written and the e-mails sent.\n\nThe report, the Verification and the evidence are those of the evaluation whose decision the\ncase holds (`evaluation_no` of the answer). It is the case's own number unless a\nre-evaluation is running or ended without a decision. A case with no report answers with\nwhat it has; an unknown `case_id` is a 404.",
        "operationId": "get_case_detail_api_v1_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Case Detail",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/feedback": {
      "get": {
        "description": "The case's feedback, oldest first.",
        "operationId": "get_feedback_api_v1_cases__case_id__feedback_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedbackAnswer"
                  },
                  "title": "Response Get Feedback Api V1 Cases  Case Id  Feedback Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Feedback",
        "tags": [
          "cases"
        ]
      },
      "post": {
        "description": "Record the operator's verdict on the case, with the session's subject as its author.\n\nThe body's `case_id` must be the one in the path, otherwise 422 and nothing is written.",
        "operationId": "post_feedback_api_v1_cases__case_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatorFeedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Post Feedback",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/v1/cases/{case_id}/steps": {
      "get": {
        "description": "The case's agent runs in start order (so evaluation by evaluation), each with its\nevaluation number and its tool calls.\n\nAn unknown `case_id` is a 404.",
        "operationId": "get_case_steps_api_v1_cases__case_id__steps_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CaseStep"
                  },
                  "title": "Response Get Case Steps Api V1 Cases  Case Id  Steps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Case Steps",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}